Routes refactor
