+++ datadrone/templates/details.html
@@ -3,7 +3,7 @@
 {% block body %}
 
 <div class="right">
-  <a href="{{ url_for('item_edit', item_id=item.item_id) }}" class="link-small">Edit</a>
+  <a href="{{ url_for('items.edit', item_id=item.item_id) }}" class="link-small">Edit</a>
 </div>
 <h3>{{ item.itemname }}</h3>
 
@@ -11,7 +11,7 @@
   Filter
 </div>
 <div class="filter-body">
-  <form method="POST" action="{{ url_for('details', item_id=item.item_id) }}">
+  <form method="POST" action="{{ url_for('items.details', item_id=item.item_id) }}">
     {{ form.hidden_tag() }}
     <div class="filter-category">
       <span class="filter-name">Date</span>
@@ -51,12 +51,12 @@
 {% endif %}
 <br>
 <select class="select-lastdays" name="lastdays" onchange="document.location.href=this.value">
-  <option value="{{ url_for('details', item_id=item.item_id, days=30) }}" {{ 'selected' if days == 30 }}>30</option>
-  <option value="{{ url_for('details', item_id=item.item_id, days=90) }}" {{ 'selected' if days == 90 }}>90</option>
-  <option value="{{ url_for('details', item_id=item.item_id, days=180) }}" {{ 'selected' if days == 180 }}>180</option>
-  <option value="{{ url_for('details', item_id=item.item_id, days=365) }}" {{ 'selected' if days == 365 }}>365</option>
+  <option value="{{ url_for('items.details', item_id=item.item_id, days=30) }}" {{ 'selected' if days == 30 }}>30</option>
+  <option value="{{ url_for('items.details', item_id=item.item_id, days=90) }}" {{ 'selected' if days == 90 }}>90</option>
+  <option value="{{ url_for('items.details', item_id=item.item_id, days=180) }}" {{ 'selected' if days == 180 }}>180</option>
+  <option value="{{ url_for('items.details', item_id=item.item_id, days=365) }}" {{ 'selected' if days == 365 }}>365</option>
 </select> days</span> |
-<a href="{{ url_for('details', item_id=item.item_id, days='all') }}" class="link-small">All</a><br>
+<a href="{{ url_for('items.details', item_id=item.item_id, days='all') }}" class="link-small">All</a><br>
 <br><br>
 
 First: {{ stats['first'].strftime('%Y-%m-%d %H:%M') if not stats['first'] == 0 else 0 }}<br>
@@ -259,7 +259,7 @@ Total today: {{ stats['total_today'] }}<br>
     {% for entry in entries[::-1] %}
       <tr>
         <td>
-          <a href="{{ url_for('entry', entry_id=entry.entry_id)}}">{{ entry["timestamp"].strftime('%Y-%m-%d %H:%M') }}</a>
+          <a href="{{ url_for('entries.entry', entry_id=entry.entry_id)}}">{{ entry["timestamp"].strftime('%Y-%m-%d %H:%M') }}</a>
         </td>
         <td>
           {% if entry["latitude"] and entry["longitude"] %}
