+++ datadrone/templates/item_edit.html
@@ -3,11 +3,11 @@
 {% block body %}
 
 <div>
-  <a href="{{ url_for('details', item_id=item.item_id) }}" class="link-small"><- Back to {{ item.itemname }}</a>
+  <a href="{{ url_for('items.details', item_id=item.item_id) }}" class="link-small"><- Back to {{ item.itemname }}</a>
 </div>
 <br>
 <div class="right">
-  <a href="{{ url_for('item_delete', item_id=item.item_id) }}" class="link-small" onClick="return confirm('Are you sure you want to delete this item?')">Delete</a>
+  <a href="{{ url_for('items.delete', item_id=item.item_id) }}" class="link-small" onClick="return confirm('Are you sure you want to delete this item?')">Delete</a>
 </div>
 <h3>Edit Item</h3>
 <form method="POST" action="">
@@ -25,20 +25,20 @@
 <h3>Tags</h3>
 {% for tag in item.tags %}
   {% if not tag.deleted %}
-  <form method="POST" action="{{ url_for('item_edittag', tag_id=tag.tag_id) }}" class="edit-tag-form">
+  <form method="POST" action="{{ url_for('tags.edit', tag_id=tag.tag_id) }}" class="edit-tag-form">
     {{ edit_tag_form.hidden_tag() }}
     {{ edit_tag_form.tagname(value=tag.name, class="form form-small") }}
     {{ edit_tag_form.hidden(checked=tag.hidden, title="hidden") }}
     {{ edit_tag_form.archived(checked=tag.archived, title="archived") }}
     {{ edit_tag_form.submit(class="btn btn-small") }}
-    <a href="{{ url_for('tag_delete', tag_id=tag.tag_id) }}" onClick="return confirm('Are you sure you want to delete this tag?')"><span class="small-text">Delete</span></a>
+    <a href="{{ url_for('tags.delete', tag_id=tag.tag_id) }}" onClick="return confirm('Are you sure you want to delete this tag?')"><span class="small-text">Delete</span></a>
   </form>
   {% endif %}
 {% endfor %}
 <br><br>
 {% endif %}
 
-<form method="POST" action="{{ url_for('item_addtag', item_id=item.item_id) }}">
+<form method="POST" action="{{ url_for('tags.add', item_id=item.item_id) }}">
 {{ add_tag_form.hidden_tag() }}
 Add tag: {{ add_tag_form.tagname(class="form form-small") }}
 {{ add_tag_form.submit(class="btn btn-small") }}
@@ -54,7 +54,7 @@ Add tag: {{ add_tag_form.tagname(class="form form-small") }}
 </form>
 <br>
 <h3>Misc</h3>
-<a href="{{ url_for('item_download', item_id=item.item_id) }}">Download csv</a>
+<a href="{{ url_for('items.download', item_id=item.item_id) }}">Download csv</a>
 
 
 {% endblock %}
