+++ datadrone/templates/list.html
@@ -5,7 +5,7 @@
 {% if items.all()|length %}
   {% for item in items %}
 
-  <form method="POST" action="{{ url_for('item_addentry', item_id=item.item_id) }}" id="add_entry_form-{{ item.item_id }}">
+  <form method="POST" action="{{ url_for('entries.add', item_id=item.item_id) }}" id="add_entry_form-{{ item.item_id }}">
     {{ entry_form.csrf_token() }}
     {{ entry_form.latitude(id="latitude-" ~ item.item_id) }}
     {{ entry_form.longitude(id="longitude-" ~ item.item_id) }}
@@ -37,7 +37,7 @@
           {% endif %}
         {% endif %}
       </span>
-      <a href="{{ url_for('details', item_id=item.item_id) }}"><div class="item-title">{{ item.itemname }}</div></a>
+      <a href="{{ url_for('items.details', item_id=item.item_id) }}"><div class="item-title">{{ item.itemname }}</div></a>
       <div class="item-spotlight">Last entry {{ spotlight_stat[item.item_id] }}</div>
     </div>
 
@@ -65,7 +65,7 @@
 <div id="add-item-modal-fade" class="modal-fade"></div>
 <div id="add-item-modal-newitem">
   What do you want to keep track of?
-  <form action="{{ url_for('item_add') }}" method="post">
+  <form action="{{ url_for('items.add') }}" method="post">
   {{ item_form.hidden_tag() }}
     <input type="submit" class="btn btn-plus" Value="Add">
     {{ item_form.itemname(class="form form-large", id="add-item-input") }}
