+++ datadrone/templates/login.html
@@ -36,6 +36,6 @@
     {{ form.submit(class="btn btn-large") }}
 </form>
 <br>
-<a href="{{ url_for('reset_request') }}" class="link-small">Forgot password?</a>
+<a href="{{ url_for('auth.reset_request') }}" class="link-small">Forgot password?</a>
 
 {% endblock%}
