+++ requirements.txt
@@ -1,12 +1,4 @@
 bcrypt==4.1.2
-blinker==1.7.0
-cffi==1.16.0
-click==8.1.7
-colorama==0.4.6
-cryptography==41.0.7
-csvalidate==1.1.2
-dnspython==2.4.2
-email-validator==2.1.0.post1
 flake8==6.1.0
 Flask==3.0.0
 Flask-Bcrypt==1.0.1
@@ -14,23 +6,8 @@ Flask-CSV==1.2.1
 Flask-Login==0.6.3
 Flask-Mail==0.9.1
 Flask-SQLAlchemy==3.1.1
-Flask-SSLify==0.1.5
 Flask-WTF==1.2.1
-greenlet==3.0.2
 gunicorn==21.2.0
-idna==3.6
 itsdangerous==2.1.2
-Jinja2==3.1.2
-MarkupSafe==2.1.3
-marshmallow==3.20.1
-packaging==23.2
-pip-review==1.3.0
 psycopg2==2.9.9
-pycparser==2.21
-PyMySQL==1.1.0
-python-dotenv==1.0.0
-six==1.16.0
-SQLAlchemy==2.0.23
-typing_extensions==4.9.0
-Werkzeug==3.0.1
-WTForms==3.1.1
+WTForms==3.1.2
